requirements.txt
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.1.0
bcrypt==4.3.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.13.0
cachetools==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.11
Flask==3.1.0
flask-cors==5.0.1
Flask-JWT-Extended==4.7.1
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
lxml==5.3.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
msgspec==0.19.0
mysql-connector-python==9.2.0
narwhals==1.26.0
numpy==2.2.2
opencv-python==4.12.0.88
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pillow==11.1.0
playwright==1.50.0
plotly==6.0.0
protobuf==5.29.3
pyarrow==19.0.0
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
PyMySQL==1.1.1
pyodbc==5.2.0
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytrends==4.9.2
pytz==2025.1
referencing==0.36.2
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
selenium==4.28.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.38
starlette==0.46.1
streamlit==1.42.0
tenacity==9.0.0
toml==0.10.2
tornado==6.4.2
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
watchdog==6.0.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0
